@86eqdp41g #Moved the lottie animation to components index
