+++ apps/web-backoffice-generic/src/components/Blocks/Account/EmailVerification/EmailVerificationForm.tsx
@@ -1,10 +1,9 @@
 import { Box } from "@mui/material";
 import { forgotPasswordSchema, ForgotPasswordType } from "./validation";
-import { Alert, TextField, Button, Link } from "core-library/components";
+import { Alert, TextField, Button, Link, LottieAnimation } from "core-library/components";
 import { FormProvider, useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { ForgotPasswordImage } from "core-library/assets";
-import { LottieAnimation } from "core-library/system/app/internal/blocks/Account/PasswordChange/components/LottieAnimation/LottieAnimation";
 
 interface Props {
   onSubmit: (values: ForgotPasswordType) => void;
