+++ packages/core-library/components/index.ts
@@ -43,6 +43,7 @@ export * from "./ToolTip/CustomToolTip";
 export * from "./PageLoader";
 export * from "./Badge/CustomBadge";
 export * from "./CSPHead";
+export * from "./LottieAnimation/LottieAnimation";
 // rows
 export * from "./table/tableRows/category/CategoryOptionsTableRow";
 export * from "./Tabs/Tabs";
